Explicitly set the recordtype on periodic events
On high optimization levels the compiler seems to not always zero out
the structure. This leads to the recordtype sometimes having a non-0
value, which may happen to be 1 (LAUNCHERAIM) which crashes the game
since the launcher doesn't exist.

If we just explicitly set it to 0, then that crash goes away.